Function has been moved to utils file
